[FEAT] Walks 데이터 fetch를 위한 react query 적용 (#8)
* [FEAT] Walks 데이터 fetch를 위한 react query 적용

* [FIX] 중복되는 type 모듈화
