+++ client/apis/api.ts
@@ -1,0 +1,5 @@
+const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
+
+export const API = {
+  WALKS: `${BASE_URL}/communities`,
+};
