+++ client/components/Navbar.tsx
@@ -36,7 +36,7 @@ export default function Navbar() {
         }
       }
     }
-    @media screen and (min-width: 768px) {
+    @media screen and (min-width: 769px) {
       display: none;
     }
   `;
