+++ client/models/WalkDefault.ts
@@ -1,0 +1,9 @@
+export interface WalkDefault {
+  projectId: number;
+  title: string;
+  time: string;
+  address: string;
+  imgUrl: string;
+  capacity: number;
+  participant: number;
+}
